requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.34.0
pydantic>=2.5.0
pytest>=7.4.0
hypothesis>=6.92.0
moto>=4.2.0
pytest-cov>=4.1.0
